Fix packaging if .manifest does not exist (Windows)

+++ metgem_packaging/tasks.py
@@ -92,8 +92,9 @@ def embed_manifest(ctx, debug):
 
     # Embed manifest in exe
     manifest = exe + '.manifest'
-    winmanifest.UpdateManifestResourcesFromXMLFile(exe, manifest)
-    os.remove(manifest)
+    if os.path.exists(manifest):
+        winmanifest.UpdateManifestResourcesFromXMLFile(exe, manifest)
+        os.remove(manifest)
 
 
 # noinspection PyShadowingNames,PyUnusedLocal
